MeetingDetail 관련 latitude, longitude 자료형 수정

+++ core/data/src/main/java/com/onmoim/core/data/model/MeetingDetail.kt
@@ -13,6 +13,6 @@ data class MeetingDetail(
     val attendance: Boolean,
     val isLightning: Boolean,
     val imgUrl: String?,
-    val latitude: Int,
-    val longitude: Int
+    val latitude: Double,
+    val longitude: Double
 )
