+++ core/network/src/main/java/com/onmoim/core/network/model/MeetingDetailDto.kt
@@ -18,9 +18,9 @@ data class MeetingDetailDto(
     @SerialName("joinCount")
     val joinCount: Int,
     @SerialName("latitude")
-    val latitude: Int,
+    val latitude: Double,
     @SerialName("longitude")
-    val longitude: Int,
+    val longitude: Double,
     @SerialName("placeName")
     val placeName: String,
     @SerialName("startDate")
